Initial attempt at molecular alignment.
By default the stack builder stacks molecules up the z-axis. This is
fine for inputted molecules which already are aligned in the xy-plane (a
default when using some molecular builders), but problematic if not. A
properly aligned molecule is also less likely to cause issues with
monomers intersecting with adjacent layers.

This script does the following:
- Moves the molecule to place its centroid at the origin

- Calculates the RMSD of each atom with the xy-plane (sums the squares
of each z-coordinate)

- Calculates the RMSD for each point for every xyz rotation across 0-180
degrees, at 20 degree intervals

- For the xyz triplet that most greatly minimises the RMSD, the step
size decreases, and the range of angles to probe in each axis is
reduced about the previously found minimising set

- This process continues, converging on the rotation that minimises the
RMSD, until the step size decreases below a certain interval

- This rotation is then applied to the inputted structure, and saved