Include alignment tool and cleanup
Inner workings of flattening/flatten.py now added to main StackBuilder,
allowing the user to use a monomer that hasn't been "cleaned up" - i.e.
been centred to the origin and made flat in the xy-plane. This is
necessary due to StackBuilder working by adding layers along the z-axis,
requiring the molecule to be flat in the orthogonal plane.

In essence this works by minimising a function which defines the RMSD of
a structure to a rotation of the xy-plane. Such a plane is the one in
which the molecule is flattest.

As well, some better commenting and cleanup of work files.